Move CTransfer_ConjugateHeatVars
